requirements.txt
torch>=2.1.0
torchvision>=0.16.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0
super-image>=0.1.7
PyYAML>=6.0
tqdm>=4.66.0
numpy>=1.24.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0
einops>=0.7.0
